Update AlgebraicConsructionOfQuasiSplitTorus.tex
Hi Armin and Eric,

I made some minor edits to the file.  It seems okay to me.  I agree with the fixed typos in my last edit: sorry about that.  I was in a hurry.  The only non-trivial edit is an additional line of explanation in Example 7 that I had in my original edit but neglected to transfer to Github.
More importantly, the additional explanation I originally believed was required from the proof of the no-name lemma, I no longer think is necessary.  I think it is sufficiently explained.